pyproject.toml
[tool.poetry]
name = "sentiric-knowledge-query-service"
version = "0.4.6"
description = "Retrieval Augmented Generation (RAG) Query Service for vector database lookups (Qdrant)."
authors = ["Sentiric Team <dev@sentiric.ai>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
structlog = "^24.4.0"
torch = {version = "2.4.1", source = "pytorch"}
sentence-transformers = "^3.1.1"
qdrant-client = "^1.12.0"
httpx = "^0.27.2"
grpcio = "^1.68.0"
grpcio-tools = "^1.68.0"
grpcio-health-checking = "^1.68.0"
prometheus-client = "^0.21.0"
protobuf = "^5.29.0"
sentiric-contracts-py = { git = "https://github.com/sentiric/sentiric-contracts.git", tag = "v1.13.6" }

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"